fix unsupported network bug & add disconnect/switch account from unsupported
