+++ package.json
@@ -189,7 +189,8 @@
         "@types/react": "17.0.2",
         "@types/react-dom": "17.0.2",
         "react-error-overlay": "6.0.9",
-        "@web3-onboard/core@^2.8.5": "patch:@web3-onboard/core@npm%3A2.8.5#./.yarn/patches/@web3-onboard-core-npm-2.8.5-10c73a002c.patch"
+        "@web3-onboard/core@^2.8.5": "patch:@web3-onboard/core@npm%3A2.8.5#./.yarn/patches/@web3-onboard-core-npm-2.8.5-10c73a002c.patch",
+        "@gooddollar/goodprotocol": "1.0.29-beta.5"
     },
     "dependencies": {
         "@gooddollar/goodprotocol": "1.0.29-beta.5",
